// src/types/PersonalHealthQuestionnaire.d.ts
export type REST_PersonalHealthQuestionnaire = {
    epilepsy: KintoneRecordField.DropDown;
    abnormalBleeding: KintoneRecordField.DropDown;
    heartProblems: KintoneRecordField.DropDown;
    shortBreath: KintoneRecordField.DropDown;
    lossVision: KintoneRecordField.DropDown;
    fatigue: KintoneRecordField.DropDown;
    ref: KintoneRecordField.Number;
    hiv: KintoneRecordField.DropDown;
    heightUnit: KintoneRecordField.RadioButton;
    height: KintoneRecordField.Number;
    headInjury: KintoneRecordField.DropDown;
    vomiting: KintoneRecordField.DropDown;
    Text_1: KintoneRecordField.SingleLineText;
    dizziness: KintoneRecordField.DropDown;
    migranes: KintoneRecordField.DropDown;
    tuberculosis: KintoneRecordField.DropDown;
    weight: KintoneRecordField.Number;
    asthma: KintoneRecordField.DropDown;
    backProblems: KintoneRecordField.DropDown;
    highBloodPressure: KintoneRecordField.DropDown;
    anemia: KintoneRecordField.DropDown;
    anxiety: KintoneRecordField.DropDown;
    brokenBones: KintoneRecordField.DropDown;
    arthritis: KintoneRecordField.DropDown;
    name: KintoneRecordField.SingleLineText;
    weightUnit: KintoneRecordField.RadioButton;
    neckPain: KintoneRecordField.DropDown;
    lossHearing: KintoneRecordField.DropDown;
    emphysema: KintoneRecordField.DropDown;
    weakness: KintoneRecordField.DropDown;
    q1: KintoneRecordField.MultiLineText;
    q2: KintoneRecordField.MultiLineText;
    q3: KintoneRecordField.MultiLineText;
    q4: KintoneRecordField.MultiLineText;
    q5: KintoneRecordField.MultiLineText;
    q6: KintoneRecordField.MultiLineText;
    fainting: KintoneRecordField.DropDown;
    hepatitis: KintoneRecordField.DropDown;
    stroke: KintoneRecordField.DropDown;
    hemorrhoids: KintoneRecordField.DropDown;
    cancer: KintoneRecordField.DropDown;
    depression: KintoneRecordField.DropDown;
    diabetes: KintoneRecordField.DropDown;
    age: KintoneRecordField.Number;

    office: KintoneRecordField.UserSelect;
};